Move classes to src hierarchy
